Response class fixes and retry logic fixes
